update home_drawer, added assets and flutter_gen, clean up code
