+++ apps/multichoice/pubspec.yaml
@@ -11,13 +11,17 @@ dependencies:
   auto_route: ^8.1.3
   bloc: ^8.1.3
   core: ^0.0.1
+  file_picker: ^8.0.5
   firebase_core: ^3.0.0
   flutter:
     sdk: flutter
   flutter_bloc: ^8.1.4
   flutter_hooks: ^0.20.5
+  flutter_svg: ^2.0.10+1
+  flutter_svg_provider: ^1.0.7
   gap: ^3.0.1
   models: ^0.0.1
+  permission_handler: ^11.3.1
   provider: ^6.1.1
   shared_preferences: ^2.2.2
   theme: ^0.0.1
@@ -29,9 +33,16 @@ dependencies:
 dev_dependencies:
   auto_route_generator: ^8.0.0
   build_runner: ^2.4.8
+  flutter_gen_runner: ^5.5.0+1
   flutter_test:
     sdk: flutter
   very_good_analysis: ^5.1.0
 
 flutter:
   uses-material-design: true
+  assets:
+    - assets/images/
+
+flutter_gen:
+  integrations:
+    flutter_svg: true
